Refactor SeatInfo to become billing info
